completed work
lesson 7 and lesson 8 complete
P.s. without admin profile

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\PostRequest;
 use App\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -10,6 +11,11 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -49,13 +55,13 @@ class PostController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(Request $request)
+    public function store(PostRequest $request)
     {
         $post = new Post();
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->desc = $request->desc;
-        $post->author_id = rand(1,4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')){
             $path = Storage::putfile('public', $request->file('img'));
@@ -79,6 +85,10 @@ class PostController extends Controller
         $post = Post::join('users', 'author_id', '=', 'users.id')
         ->find($id);
 
+        if (!$post){
+            return redirect()->route('post.index')->withErrors('Произошла непредвиденная (предвиденная) ошибка. Код ошибки: BRUH1337');
+        }
+
         return view('posts.show', compact('post'));
     }
 
@@ -91,6 +101,14 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (!$post){
+            return redirect()->route('post.index')->withErrors('Произошла непредвиденная (предвиденная) ошибка. Код ошибки: BRUH360');
+        }
+
+        if ($post->author_id !=\Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
@@ -101,9 +119,16 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(Request $request, $id)
+    public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+        if (!$post){
+            return redirect()->route('post.index')->withErrors('Произошла непредвиденная (предвиденная) ошибка. Код ошибки: BRUH420');
+        }
+
+        if ($post->author_id !=\Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
 
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
@@ -129,6 +154,14 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (!$post){
+            return redirect()->route('post.index')->withErrors('Произошла непредвиденная (предвиденная) ошибка. Код ошибки: BRUH1234');
+        }
+
+        if ($post->author_id !=\Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('Вы не можете удалить данный пост');
+        }
+
         $post->delete();
         return redirect()->route('post.index')->with('success', 'Пост успешно удален');
     }
